# pytorrent/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AppStatusResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiOk"
          },
          {
            "properties": {
              "status": {
                "$ref": "#/components/schemas/AppStatus"
              }
            },
            "required": [
              "status"
            ],
            "type": "object"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "description": "Optional per-user API token. Tokens are accepted only when PYTORRENT_AUTH_ENABLE=1. When auth is disabled, API endpoints do not require credentials.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "bearerApiToken": {
        "description": "Optional per-user API token in Authorization: Bearer <token>. Tokens are accepted only when PYTORRENT_AUTH_ENABLE=1.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Optional Flask session cookie. Required only when PYTORRENT_AUTH_ENABLE=1 and no API token is used.",
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      }
    },
    "parameters": {
      "ProfileId": {
        "name": "profile_id",
        "in": "query",
        "required": false,
        "description": "Optional rTorrent profile id for profile-scoped API operations. If omitted, the user active profile is used for backward compatibility; fallback profile id 1 is used only when no active profile exists and the user can access it.",
        "schema": {
          "type": "integer",
          "default": 1,
          "minimum": 1
        }
      },
...
    }
  },
  "info": {
    "title": "pyTorrent API",
    "version": "0.0.1"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/app/status": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppStatusResponse"
                }
              }
            },
            "description": "Application status"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "pyTorrent application status",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
...
    "/api/cleanup/all": {
      "post": {
        "description": "Clears finished job logs, Smart Queue history, active-profile operation logs, Planner action history and automation history. Pending/running jobs, saved rules, settings and torrents are preserved.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            },
            "description": "Cleanup result"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear all cleanup-supported history",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/cleanup/automations": {
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            },
            "description": "Cleanup result"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear automation execution history",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/cleanup/cache": {
      "post": {
        "description": "Clears active profile runtime and database-backed caches without deleting torrents, settings, rules or logs.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            },
            "description": "Cleanup result"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No active profile"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear active profile cache",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/cleanup/database/vacuum": {
      "post": {
        "summary": "Vacuum application database",
        "description": "Runs SQLite database vacuum/maintenance for admin users. Note: the cleanup summary is returned with the result.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "force": {
                    "type": "boolean"
                  },
                  "profile_id": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1,
                    "description": "Optional rTorrent profile id. Overrides active profile for this request."
                  },
                  "profile_name": {
                    "type": "string",
                    "description": "Optional rTorrent profile name. Used only when profile_id is not provided."
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "vacuum": {
                          "type": "object",
                          "additionalProperties": true
                        },
                        "cleanup": {
                          "$ref": "#/components/schemas/CleanupSummary"
                        }
                      }
                    }
                  ]
                }
              }
            },
            "description": "Vacuum result"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/cleanup/jobs": {
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            },
            "description": "Cleanup result"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear finished job history",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/cleanup/operation-logs": {
      "post": {
        "description": "Clears active-profile operation logs. Torrents, jobs, settings and rules are preserved.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            },
            "description": "Cleanup result"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear operation logs",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/cleanup/planner": {
      "post": {
        "description": "Deletes Download Planner action history for the active profile. Saved Planner settings are preserved.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            },
            "description": "Cleanup result"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No active profile"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear Planner action history",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/cleanup/poller-diagnostics": {
      "post": {
        "summary": "Reset poller diagnostics",
        "description": "Clears in-memory poller runtime counters for the active profile. Note: polling settings and torrent state are preserved.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CleanupResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "runtime": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      }
                    }
                  ]
                }
              }
            },
            "description": "Cleanup result"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/cleanup/smart-queue": {
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            },
            "description": "Cleanup result"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear Smart Queue history",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/cleanup/summary": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            },
            "description": "Cleanup summary"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Cleanup summary",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
...
    "/api/jobs/clear": {
      "post": {
        "description": "Deletes jobs that are not pending or running.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "deleted": {
                          "type": "integer"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear finished job history",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
...
  }
}